Fix initialization when not running inside Passenger
The Union Station hooks are not supposed to initialize when running
outside Passenger. But they did so anyway, probably because I accidentally
made it that way in order to be able to test the code inside the test
suite (which partially runs without Passenger).

The problem is only triggered if the user has union_station_hooks_core
in his Gemfile, AND he tries to run Rails command (e.g. 'rails console')
outside Passenger.